+++ wangcai_iOS/wangcai/view_iPhone/common/BaseTaskTableViewController.h
@@ -50,6 +50,6 @@
 - (void)enableTixianButton:(BOOL)enabled;
 
 //带动画和声音设置余额
-- (void)setYuENumberWithAnimationFrom:(float)oldNum toNum:(float)newNum;
+- (void)setYuENumberWithAnimationFrom:(int)oldNum toNum:(int)newNum;
 
 @end
